requirements.txt
opencv-contrib-python
python-dotenv